Adding NonLinearSvm classes
Adding non-linear SVMs and fixing the objective function calculation for
those types of classes. Added IRIS test data for non-linearly separable
test data and added a unit test for non-linear SVMs. Accuracy seems to
be better with the Linear SVM on the non-linearly separable data so far;
more research should clear up the cause of this.